Improve character role selector
