+++ src/main/python/plotlyst/view/character_editor.py
@@ -45,7 +45,7 @@ from src.main.python.plotlyst.view.widget.big_five import BigFivePersonalityWidg
 from src.main.python.plotlyst.view.widget.character.editor import CharacterAgeEditor
 from src.main.python.plotlyst.view.widget.character.plan import CharacterPlansWidget
 from src.main.python.plotlyst.view.widget.character.topic import CharacterTopicsEditor
-from src.main.python.plotlyst.view.widget.characters import CharacterRoleSelector
+from src.main.python.plotlyst.view.widget.character.editor import CharacterRoleSelector
 from src.main.python.plotlyst.view.widget.template import CharacterProfileTemplateView
